llm_eval_framework/requirements.txt
# Core dependencies for LLM Evaluation Framework
pydantic>=2.0.0
click>=8.0.0
pyyaml>=6.0.0
jinja2>=3.0.0
aiohttp>=3.8.0
litellm>=1.0.0
rich>=13.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-multipart>=0.0.6